requirements.txt
ansible==1.9.2
apt-xapian-index==0.44
awscli==1.8.3
boto==2.38.0
boto3==1.1.3
botocore==1.2.1
cffi==1.2.1
chardet==2.0.1
colorama==0.3.3
command-not-found==0.2.44
cryptography==1.0.1
docutils==0.12
ecdsa==0.13
enum34==1.0.4
futures==2.2.0
GnuPGInterface==0.3.2
httplib2==0.9.1
idna==2.0
ipaddress==1.0.14
Jinja2==2.8
jmespath==0.7.1
language-selector==0.1
MarkupSafe==0.23
ndg-httpsclient==0.4.0
paramiko==1.15.2
pyasn1==0.1.8
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==0.15.1
python-apt===0.8.3ubuntu7
python-dateutil==2.4.2
python-debian===0.1.21ubuntu1
python-digitalocean==1.6
PyYAML==3.11
requests==2.7.0
rsa==3.1.4
six==1.9.0
ufw==0.31.1.post1
urllib3==1.12
